Fix linting issues

Closes #62

See merge request feng/training/material!54

+++ courses/fundamentals_of_ada/mini_projects/cinema/package.sh
@@ -1,13 +1,13 @@
 #! /usr/bin/env bash
 set -ex
 
-root=$(dirname $(readlink -e "$0"))
-out="$root/pkg/$(basename $root)"
+root=$(dirname "$(readlink -e "$0")")
+out="$root/pkg/$(basename "$root")"
 zip="$1"
 test ! -z "$zip"
 
 test -d "$out" || mkdir -p "$out"
-rm -rf "$out/*"
+rm -rf "${out:?}/*"
 
 for sub in src answer mini_cinema.gpr mini_cinema.adc include logs resources tests testsuite README.md; do
     test -f "$root/$sub" -o -d "$root/$sub"
@@ -15,8 +15,8 @@ for sub in src answer mini_cinema.gpr mini_cinema.adc include logs resources tes
 done
 
 (
-cd $(dirname $out)
+cd "$(dirname "$out")"
 
 rm -f "$zip" 2>/dev/null
-zip -r "$zip" "$(basename $out)"
+zip -r "$zip" "$(basename "$out")"
 )
